requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
typing-extensions>=4.5.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.1

# Document Processing (Simplified)
python-docx==1.0.1
PyPDF2==3.0.1
python-magic>=0.4.27
python-magic-bin>=0.4.14
reportlab==4.0.7

# Data Processing (Simplified)
pandas==2.1.3
numpy==1.26.2

# Database
SQLAlchemy==2.0.27
alembic==1.13.1

# Async Support
aiofiles==23.2.1
aiohttp>=3.8.0

# Utilities
PyYAML>=6.0
requests>=2.31.0
certifi>=2023.7.22
python-json-logger>=2.0.0

# Text Processing (Simplified)
beautifulsoup4==4.12.3
unidecode==1.3.8
contractions==0.1.73
scikit-learn==1.3.2
sentence-transformers==2.2.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Web Framework
jinja2>=3.0.0
